Export headings and desktop and add visiblity prop
- Export the array of headings
- Export if desktop is active or not
- Add option to disable visibility

Knowing if Toc is in desktop mode enables me to change the layout according to that and change my layout to use the space without having to make extra calculations

I want to hide Toc when there are less than 2 headlines. So the headlines are exported to be able to count them and a visible prop is added to be able to hide the Toc.