improve readme doc on CSS class toc-exclude in default heading selector
triggered by question in #15
also fix Svelte 3.52.0 warnings

Toc.svelte:118:12 A11y: visible, non-interactive elements with an on:click event must be accompanied by an on:keydown, on:keyup, or on:keypress event.
Toc.svelte:118:12 A11y: visible, non-interactive elements with an on:click event must be accompanied by an on:keydown, on:keyup, or on:keypress event.